requirements.txt
flask
google-auth-oauthlib
google-api-python-client
moviepy
SpeechRecognition
pydub
opencv-python
numpy
transformers
torch
google-cloud-translate==2.0.1